Dark Mode
Dark Mode theming and fixing the navbar so it doesn't overlap the header when scrolling.
Todo:
twitter embed needs to be set to dark mode when changed.  to be done via script not css
SVG icons in the contact us need a class or an id so we can invert them in dark theme
FAQ needs to be properly themed (right now it's just inverted)
shadows could use some tweaking